122 auth middleware package (#157)
* auth-middleware library (package) + test project for demonstration of usage

added auth-middleware library (package), and a test project for demonstration of usage

* improve docs for running test svc

* expect id_token to be in Authorization header

* phone-verify: use env variables; improve handling of decode errors

use environment variables for phone-verify-secret and phone-verify-audience; improve handling of decode errors

* comment google authentication in middleware since we're no longer using it

comment google auth to indicate that we're no longer using it.  I'm not deleting it in case someone wants to look at it later.  Feel free to delete if necessary

* only check issuer with 'https://' + oidc_backend_hostname

don't check issuer (iss) against raw hostname, but only check issuer (iss) against oidc backend hostname prepended with 'https://'

* docs on the env variables needed to use the auth-middleware package

documentation on the environment variables that are needed to use the auth-middleware package

* Update .gitignore in include Python build directories.

* Updated docker compose YAML file and README to include environment variable file.

* Add cryptography package to fix runtime error.

* Use signature verification and use OIDC client ID.

* Update auth middleware library README.

* Update CHANGELOG.